# api/openapi.yaml
openapi: 3.0.3
info:
  title: GoAuction API
  version: 1.0.0
  description: |
    Real-time auction engine. Concurrent bidding with optimistic locking,
    anti-sniping, JWT auth, WebSocket updates per lot.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local docker-compose
...
paths:
...
  /api/lots:
    get:
      tags: [lots]
      summary: List lots (paginated, optional status filter)
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: status
          schema: { type: string, enum: [active, closed, canceled] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Lot' }
    post:
      tags: [lots]
      summary: Create a new lot (auth required)
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateLotRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lot' }
        '400': { description: Invalid lot parameters }
        '401': { description: Unauthorized }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Lot:
      type: object
      properties:
        id:             { type: string, format: uuid }
        title:          { type: string }
        start_price:    { type: number, format: double }
        min_step:       { type: number, format: double }
        current_price:  { type: number, format: double }
        status:         { type: string, enum: [active, closed, canceled] }
        created_at:     { type: string, format: date-time }
        closing_at:     { type: string, format: date-time }
        version:        { type: integer }
        winner_id:      { type: string, format: uuid, nullable: true }
        extended_count: { type: integer }
    CreateLotRequest:
      type: object
      required: [title, start_price, min_step, closing_at]
      properties:
        title:       { type: string }
        start_price: { type: number, format: double }
        min_step:    { type: number, format: double }
        closing_at:  { type: string, format: date-time }
...